OPENEUROPA-1963: Move namespace for entity interfaces.
